requirements.txt
streamlit>=1.32
supabase>=2.4
pandas
python-dotenv
jinja2>=3.1
playwright>=1.42
python-docx>=1.1
anthropic>=0.50
google-generativeai>=0.5
pillow>=10.0
